requirements.txt
faster-whisper
llama-cpp-python
piper-tts
sounddevice
numpy
python-dotenv
pvporcupine
numpy
pyaudio
requests
faster-whisper
piper-tts
ollama
